// MCJPortal/ClientApp/src/app/swagger-services/api.client.generated.ts
export class CountryViewModel implements ICountryViewModel {
    id!: number;
    name?: string | null;
    fileDoc?: string | null;

    constructor(data?: ICountryViewModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"] !== undefined ? _data["id"] : <any>null;
            this.name = _data["name"] !== undefined ? _data["name"] : <any>null;
            this.fileDoc = _data["fileDoc"] !== undefined ? _data["fileDoc"] : <any>null;
        }
    }

    static fromJS(data: any): CountryViewModel {
        data = typeof data === 'object' ? data : {};
        let result = new CountryViewModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id !== undefined ? this.id : <any>null;
        data["name"] = this.name !== undefined ? this.name : <any>null;
        data["fileDoc"] = this.fileDoc !== undefined ? this.fileDoc : <any>null;
        return data; 
    }
}

export interface ICountryViewModel {
    id: number;
    name?: string | null;
    fileDoc?: string | null;
}

export class ListItemViewModel implements IListItemViewModel {
    id!: string;
    listTypeId!: ListTypeEnum;
    value?: string | null;
    permissions?: RolePermissionViewModel[] | null;

    constructor(data?: IListItemViewModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"] !== undefined ? _data["id"] : <any>null;
            this.listTypeId = _data["listTypeId"] !== undefined ? _data["listTypeId"] : <any>null;
            this.value = _data["value"] !== undefined ? _data["value"] : <any>null;
            if (Array.isArray(_data["permissions"])) {
                this.permissions = [] as any;
                for (let item of _data["permissions"])
                    this.permissions!.push(RolePermissionViewModel.fromJS(item));
            }
        }
    }

    static fromJS(data: any): ListItemViewModel {
        data = typeof data === 'object' ? data : {};
        let result = new ListItemViewModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id !== undefined ? this.id : <any>null;
        data["listTypeId"] = this.listTypeId !== undefined ? this.listTypeId : <any>null;
        data["value"] = this.value !== undefined ? this.value : <any>null;
        if (Array.isArray(this.permissions)) {
            data["permissions"] = [];
            for (let item of this.permissions)
                data["permissions"].push(item.toJSON());
        }
        return data; 
    }
}

export interface IListItemViewModel {
    id: string;
    listTypeId: ListTypeEnum;
    value?: string | null;
    permissions?: RolePermissionViewModel[] | null;
}

export enum ListTypeEnum {
    Staff = 1,
    LineStatus = 2,
    ProductGroup = 3,
    FileType = 4,
    Currency = 5,
    JobStatus = 6,
    ProductionType = 7,
    Suppliers = 8,
    Permission = 12,
    Material = 13,
}

export class RolePermissionViewModel implements IRolePermissionViewModel {
    roleId?: string | null;
    permissionId!: string;
    access!: AccessEnum;

    constructor(data?: IRolePermissionViewModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.roleId = _data["roleId"] !== undefined ? _data["roleId"] : <any>null;
            this.permissionId = _data["permissionId"] !== undefined ? _data["permissionId"] : <any>null;
            this.access = _data["access"] !== undefined ? _data["access"] : <any>null;
        }
    }

    static fromJS(data: any): RolePermissionViewModel {
        data = typeof data === 'object' ? data : {};
        let result = new RolePermissionViewModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["roleId"] = this.roleId !== undefined ? this.roleId : <any>null;
        data["permissionId"] = this.permissionId !== undefined ? this.permissionId : <any>null;
        data["access"] = this.access !== undefined ? this.access : <any>null;
        return data; 
    }
}

export interface IRolePermissionViewModel {
    roleId?: string | null;
    permissionId: string;
    access: AccessEnum;
}

export enum AccessEnum {
    NoAccess = 0,
    View = 1,
    Edit = 2,
}
// ...
export class ProjectViewModel implements IProjectViewModel {
    id!: number;
    name?: string | null;
    projectLines?: ProjectLineViewModel[] | null;

    constructor(data?: IProjectViewModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"] !== undefined ? _data["id"] : <any>null;
            this.name = _data["name"] !== undefined ? _data["name"] : <any>null;
            if (Array.isArray(_data["projectLines"])) {
                this.projectLines = [] as any;
                for (let item of _data["projectLines"])
                    this.projectLines!.push(ProjectLineViewModel.fromJS(item));
            }
        }
    }

    static fromJS(data: any): ProjectViewModel {
        data = typeof data === 'object' ? data : {};
        let result = new ProjectViewModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id !== undefined ? this.id : <any>null;
        data["name"] = this.name !== undefined ? this.name : <any>null;
        if (Array.isArray(this.projectLines)) {
            data["projectLines"] = [];
            for (let item of this.projectLines)
                data["projectLines"].push(item.toJSON());
        }
        return data; 
    }
}

export interface IProjectViewModel {
    id: number;
    name?: string | null;
    projectLines?: ProjectLineViewModel[] | null;
}

export class RoleViewModel implements IRoleViewModel {
    id?: string | null;
    name?: string | null;
    normalizedName?: string | null;

    constructor(data?: IRoleViewModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"] !== undefined ? _data["id"] : <any>null;
            this.name = _data["name"] !== undefined ? _data["name"] : <any>null;
            this.normalizedName = _data["normalizedName"] !== undefined ? _data["normalizedName"] : <any>null;
        }
    }

    static fromJS(data: any): RoleViewModel {
        data = typeof data === 'object' ? data : {};
        let result = new RoleViewModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id !== undefined ? this.id : <any>null;
        data["name"] = this.name !== undefined ? this.name : <any>null;
        data["normalizedName"] = this.normalizedName !== undefined ? this.normalizedName : <any>null;
        return data; 
    }
}

export interface IRoleViewModel {
    id?: string | null;
    name?: string | null;
    normalizedName?: string | null;
}
// ...
export class UserViewModel implements IUserViewModel {
    id?: string | null;
    email!: string;
    firstName!: string;
    lastName!: string;
    userName!: string;
    password?: string | null;
    phoneNumber?: string | null;
    isActive!: boolean;
    isAllProjectsAllowed!: boolean;
    countryId?: number | null;
    country?: CountryViewModel | null;
    roleId!: string;
    userProjectIds?: number[] | null;

    constructor(data?: IUserViewModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"] !== undefined ? _data["id"] : <any>null;
            this.email = _data["email"] !== undefined ? _data["email"] : <any>null;
            this.firstName = _data["firstName"] !== undefined ? _data["firstName"] : <any>null;
            this.lastName = _data["lastName"] !== undefined ? _data["lastName"] : <any>null;
            this.userName = _data["userName"] !== undefined ? _data["userName"] : <any>null;
            this.password = _data["password"] !== undefined ? _data["password"] : <any>null;
            this.phoneNumber = _data["phoneNumber"] !== undefined ? _data["phoneNumber"] : <any>null;
            this.isActive = _data["isActive"] !== undefined ? _data["isActive"] : <any>null;
            this.isAllProjectsAllowed = _data["isAllProjectsAllowed"] !== undefined ? _data["isAllProjectsAllowed"] : <any>null;
            this.countryId = _data["countryId"] !== undefined ? _data["countryId"] : <any>null;
            this.country = _data["country"] ? CountryViewModel.fromJS(_data["country"]) : <any>null;
            this.roleId = _data["roleId"] !== undefined ? _data["roleId"] : <any>null;
            if (Array.isArray(_data["userProjectIds"])) {
                this.userProjectIds = [] as any;
                for (let item of _data["userProjectIds"])
                    this.userProjectIds!.push(item);
            }
        }
    }

    static fromJS(data: any): UserViewModel {
        data = typeof data === 'object' ? data : {};
        let result = new UserViewModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id !== undefined ? this.id : <any>null;
        data["email"] = this.email !== undefined ? this.email : <any>null;
        data["firstName"] = this.firstName !== undefined ? this.firstName : <any>null;
        data["lastName"] = this.lastName !== undefined ? this.lastName : <any>null;
        data["userName"] = this.userName !== undefined ? this.userName : <any>null;
        data["password"] = this.password !== undefined ? this.password : <any>null;
        data["phoneNumber"] = this.phoneNumber !== undefined ? this.phoneNumber : <any>null;
        data["isActive"] = this.isActive !== undefined ? this.isActive : <any>null;
        data["isAllProjectsAllowed"] = this.isAllProjectsAllowed !== undefined ? this.isAllProjectsAllowed : <any>null;
        data["countryId"] = this.countryId !== undefined ? this.countryId : <any>null;
        data["country"] = this.country ? this.country.toJSON() : <any>null;
        data["roleId"] = this.roleId !== undefined ? this.roleId : <any>null;
        if (Array.isArray(this.userProjectIds)) {
            data["userProjectIds"] = [];
            for (let item of this.userProjectIds)
                data["userProjectIds"].push(item);
        }
        return data; 
    }
}

export interface IUserViewModel {
    id?: string | null;
    email: string;
    firstName: string;
    lastName: string;
    userName: string;
    password?: string | null;
    phoneNumber?: string | null;
    isActive: boolean;
    isAllProjectsAllowed: boolean;
    countryId?: number | null;
    country?: CountryViewModel | null;
    roleId: string;
    userProjectIds?: number[] | null;
}